Deploy to Heroku via Docker
```
heroku stack:set container
```
Updates the .env file to be able to pass some environment values to
reproduce Heroku behavior in local when needed.
Note that the release.command didn't work and had to `collectstatic`
within the `run.web` of the `heroku.yml`.
Adds `curl` to the dependencies for streaming logs:
https://devcenter.heroku.com/articles/build-docker-images-heroku-yml